+++ client/src/extension.ts
@@ -13,6 +13,8 @@ import { MarkPreview, getWebviewOptions } from './preview';
 import { FoldingRangeProvider } from './providers/FoldingRangeProvider';
 import { DocumentSymbolProvider, TreeSymbol, ParagraphTreeSymbol } from './providers/DocumentSymbolProvider';
 import { SemanticTokenLegend, SemanticTokenProvider } from './providers/SemanticTokenProvider';
+import { WikiCodeActionProvider } from './providers/CodeActionProvider';
+import { TableSnippetProvider } from './providers/CompletionProvider';
 
 let client: LanguageClient;
 let activeRules: vscode.Disposable[] = [];
@@ -246,43 +248,6 @@ export async function activate(context: ExtensionContext) {
     )
   );
 
-  class TableSnippetProvider implements vscode.CompletionItemProvider {
-    provideCompletionItems(
-      document: vscode.TextDocument,
-      position: vscode.Position,
-      token: vscode.CancellationToken,
-      context: vscode.CompletionContext
-    ): vscode.ProviderResult<vscode.CompletionItem[] | vscode.CompletionList> {
-      const linePrefix = document.lineAt(position).text.substring(0, position.character);
-      const match = linePrefix.match(/(?:^|\s)table(\d{1,2})\*(\d{1,2})$/);
-
-      if (match) {
-        const rows = parseInt(match[1]);
-        const cols = parseInt(match[2]);
-
-        let snippetText = "";
-        let tabStopIndex = 1;
-        for (let r = 0; r < rows; r++) {
-          for (let c = 0; c < cols; c++) {
-            snippetText += `|| \${${tabStopIndex++}:내용} `;
-          }
-          snippetText += "||\n";
-        }
-
-        const item = new vscode.CompletionItem(`table${rows}*${cols}`, vscode.CompletionItemKind.Snippet);
-        item.insertText = new vscode.SnippetString(snippetText);
-        item.detail = `${rows}행 ${cols}열 표 삽입`;
-
-        const matchText = match[0].trimStart();
-        const matchStart = position.character - matchText.length;
-        item.range = new vscode.Range(position.line, matchStart, position.line, position.character);
-
-        return [item];
-      }
-      return undefined;
-    }
-  }
-
   context.subscriptions.push(
     vscode.languages.registerCompletionItemProvider(
       { language: 'namu' },
@@ -290,87 +255,6 @@ export async function activate(context: ExtensionContext) {
       '0', '1', '2', '3', '4', '5', '6', '7', '8', '9'
     )
   );
-
-  class WikiCodeActionProvider implements vscode.CodeActionProvider {
-    provideCodeActions(
-      document: vscode.TextDocument,
-      range: vscode.Range | vscode.Selection,
-      context: vscode.CodeActionContext,
-      token: vscode.CancellationToken
-    ): vscode.ProviderResult<(vscode.Command | vscode.CodeAction)[]> {
-      const actions: vscode.CodeAction[] = [];
-      const text = document.getText();
-      const cursorOffset = document.offsetAt(range.start);
-
-      // Code Action: Wrap
-      if (!range.isEmpty) {
-        const selected = document.getText(range);
-        [
-          { label: 'wiki', code: `{{{#!wiki \n${selected}\n}}}` },
-          { label: 'style', code: `{{{#!style\n${selected}\n}}}` },
-          { label: 'folding', code: `{{{#!folding \n${selected}\n}}}` },
-          { label: 'if', code: `{{{#!if \n${selected}\n}}}` },
-          { label: 'syntax', code: `{{{#!syntax \n${selected}\n}}}` },
-          { label: '삼중괄호', code: `{{{${selected}}}}` }
-        ].forEach(({ label, code }) => {
-          const action = new vscode.CodeAction(
-            `${label} 구문으로 선택 영역 감싸기`,
-            vscode.CodeActionKind.RefactorRewrite,
-          );
-          action.edit = new vscode.WorkspaceEdit();
-          action.edit.replace(document.uri, range, code);
-          actions.push(action);
-        });
-      }
-
-      // Code Action: Unwrap
-      const tokenRegex = /\{\{\{|\}\}\}/g;
-      const stack: number[] = [];
-      let match, foundBlock: { start: number; end: number } | null = null;
-      let minLength = Infinity;
-      
-      // Block 찾기
-      while ((match = tokenRegex.exec(text)) !== null) {
-        if (match[0] === '{{{') {
-          stack.push(match.index);
-        } else if (stack.length > 0) {
-          const start = stack.pop()!;
-          const end = match.index + 3;
-          if (start <= cursorOffset && cursorOffset <= end && end - start < minLength) {
-            minLength = end - start;
-            foundBlock = { start, end };
-          }
-        }
-      }
-      
-      // Block 해제
-      if (foundBlock) {
-        const { start, end } = foundBlock;
-        const blockText = text.substring(start, end);
-        
-        const startRegex = /^\{\{\{(?:#!([^\r\n]*)\r?\n|([+\-#][^\s]*)\s?|)/;
-        const kindMatch = blockText.match(startRegex);
-        
-        let kind = "삼중괄호";
-        if (kindMatch) {
-          if (kindMatch[1]) kind = kindMatch[1].split(' ')[0];
-          else if (kindMatch[2]) kind = kindMatch[2];
-        }
-
-        const unwrapped = blockText
-          .replace(startRegex, '')
-          .replace(/(?:\r?\n[ \t]*)?\}\}\}$/, '');
-
-        const action = new vscode.CodeAction(`현재 위치한 ${kind} 구문 벗기기`, vscode.CodeActionKind.RefactorRewrite);
-        action.edit = new vscode.WorkspaceEdit();
-        action.edit.replace(document.uri, new vscode.Range(document.positionAt(start), document.positionAt(end)), unwrapped);
-
-        actions.push(action);
-      }
-
-      return actions;
-    }
-  }
 
   context.subscriptions.push(
     vscode.languages.registerCodeActionsProvider(
